# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Electronic signature API
  version: 0.0.0
  description: >
    Public HTTP surface. Health, account-user authentication, document ingestion,
    preparation, send, and the signer-facing session API.
    This document is not a legal or compliance statement.
servers:
  - url: http://localhost:4000
paths:
...
  /organizations/{organizationId}/documents:
    post:
      summary: Create a draft document and a short-lived upload capability
      operationId: createDraftDocument
      description: >
        Authenticated organization members with document.write create a DRAFT
        document. Inspection starts pending. The response includes a short-lived
        upload URL and a one-time token. The token is not stored in idempotent
        replays. Object-storage keys are never returned.
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Draft created with an upload capability
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDocumentResponse'
        '400':
          description: Invalid title, filename, or idempotency key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: CSRF rejected, wrong organization, or role denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /organizations/{organizationId}/documents/{documentId}:
    get:
      summary: Load a tenant document without object-storage keys
      operationId: getOrganizationDocument
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDocument'
        '403':
          description: No membership in this organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Document not in this organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /organizations/{organizationId}/documents/{documentId}/source:
    put:
      summary: Stream a PDF into the issued upload capability
      operationId: uploadDocumentSource
      description: >
        PDF only. The API validates declared content type, sanitized .pdf
        extension from create time, magic bytes, and DOCUMENT_MAX_UPLOAD_BYTES.
        Reverse proxies must use the same limit. Object storage also enforces
        the cap. The original filename is stored only as sanitized display metadata.
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: x-upload-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/pdf:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Source revision stored; inspection pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDocument'
        '400':
          description: Malformed content type, extension, or magic bytes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          description: Body exceeds DOCUMENT_MAX_UPLOAD_BYTES
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /organizations/{organizationId}/documents/{documentId}/previews:
    post:
      summary: Issue a short-lived preview download capability
      operationId: issueDocumentPreview
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Preview grant issued. Object keys are not included.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuePreviewResponse'
  /document-previews/{grantId}:
    get:
      summary: Stream an authorized PDF preview
      operationId: streamDocumentPreview
      parameters:
        - name: grantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: x-preview-token
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF bytes
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          description: Missing, unknown, or expired preview token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /organizations/{organizationId}/documents/{documentId}/signers:
    put:
      summary: Replace signers and signing mode
      operationId: replaceDocumentSigners
      description: >
        Draft or prepared documents only. Ordered mode requires consecutive
        unique routingOrder values starting at 1. Parallel mode requires every
        routingOrder to be 1. Cross-tenant membership is denied.
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceSignersRequest'
      responses:
        '200':
          description: Signers replaced. May transition draft to prepared.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDocument'
        '403':
          description: Wrong organization or role denied
        '409':
          description: Preparation is locked after send
  /organizations/{organizationId}/documents/{documentId}/fields:
    put:
      summary: Replace server-owned signature fields
      operationId: replaceDocumentFields
      description: >
        Field ids are generated by the server. Client coordinates are stored
        only after page and overlap validation. Types are signature, initials,
        date_signed, and signer_name. Sent envelopes cannot have fields moved.
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceFieldsRequest'
      responses:
        '200':
          description: Fields replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDocument'
        '409':
          description: Document already sent or later
  /organizations/{organizationId}/documents/{documentId}/send:
    post:
      summary: Send a prepared document
      operationId: sendDocument
      description: >
        Transitions draft or prepared to sent. Creates one issued signing
        session per signer. Stores only token hashes. Raw tokens appear only
        in the first response and the notifier. Idempotent with Idempotency-Key.
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendDocumentRequest'
      responses:
        '200':
          description: Document sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendDocumentResponse'
  /organizations/{organizationId}/documents/{documentId}/signers/{signerId}/sessions/rotate:
    post:
      summary: Revoke open sessions and issue a new signing token
      operationId: rotateSigningSession
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: signerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: New session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateSessionResponse'
  /organizations/{organizationId}/documents/{documentId}/sessions/{sessionId}/revoke:
    post:
      summary: Revoke an issued or active signing session
      operationId: revokeSigningSession
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session revoked
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, correlationId]
          properties:
            code:
              type: string
              enum:
                - validation
                - authentication
                - not_found
                - forbidden
                - conflict
                - idempotency_replay
                - payload_too_large
                - rate_limited
                - not_ready
                - external_service
                - internal
            message:
              type: string
            correlationId:
              type: string
...
    CreateDocumentRequest:
      type: object
      additionalProperties: false
      required: [title, filename]
      properties:
        title:
          type: string
        filename:
          type: string
          description: Untrusted. Stored only after sanitization as display metadata.
    PublicDocument:
      type: object
      additionalProperties: false
      required:
        [
          documentId,
          title,
          state,
          inspectionStatus,
          displayName,
          availableForSigning,
          currentRevision,
          signingMode,
          signers,
          fields,
        ]
      properties:
        documentId:
          type: string
          format: uuid
        title:
          type: string
        state:
          type: string
        inspectionStatus:
          type: string
          enum: [pending, accepted, rejected]
        displayName:
          type: string
          nullable: true
        availableForSigning:
          type: boolean
        currentRevision:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/PublicDocumentRevision'
        signingMode:
          type: string
          enum: [ordered, parallel]
        signers:
          type: array
          items:
            $ref: '#/components/schemas/PublicSigner'
        fields:
          type: array
          items:
            $ref: '#/components/schemas/PublicSignatureField'
    PublicDocumentRevision:
      type: object
      additionalProperties: false
      required: [revisionId, contentType, sizeBytes, sha256Digest, displayName, pageCount]
      properties:
        revisionId:
          type: string
          format: uuid
        contentType:
          type: string
          enum: [application/pdf]
        sizeBytes:
          type: integer
        sha256Digest:
          type: string
        displayName:
          type: string
        pageCount:
          type: integer
    CreateDocumentResponse:
      allOf:
        - $ref: '#/components/schemas/PublicDocument'
        - type: object
          additionalProperties: false
          required: [upload]
          properties:
            upload:
              type: object
              additionalProperties: false
              required: [url, method, expiresAt, maxBytes, contentType, tokenHeader, token]
              properties:
                url:
                  type: string
                method:
                  type: string
                  enum: [PUT]
                expiresAt:
                  type: string
                  format: date-time
                maxBytes:
                  type: integer
                contentType:
                  type: string
                  enum: [application/pdf]
                tokenHeader:
                  type: string
                token:
                  type: string
                  nullable: true
    IssuePreviewResponse:
      type: object
      additionalProperties: false
      required: [url, expiresAt, tokenHeader, token, contentType]
      properties:
        url:
          type: string
        expiresAt:
          type: string
          format: date-time
        tokenHeader:
          type: string
        token:
          type: string
        contentType:
          type: string
          enum: [application/pdf]
    PublicSigner:
      type: object
      additionalProperties: false
      required: [signerId, email, displayName, routingOrder, status]
      properties:
        signerId:
          type: string
          format: uuid
        email:
          type: string
          nullable: true
        displayName:
          type: string
        routingOrder:
          type: integer
        status:
          type: string
          enum: [pending, signed, declined]
    PublicSignatureField:
      type: object
      additionalProperties: false
      required: [fieldId, signerId, type, pageNumber, x, y, width, height, required]
      properties:
        fieldId:
          type: string
          format: uuid
        signerId:
          type: string
          format: uuid
        type:
          type: string
          enum: [signature, initials, date_signed, signer_name]
        pageNumber:
          type: integer
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        required:
          type: boolean
    ReplaceSignersRequest:
      type: object
      additionalProperties: false
      required: [signingMode, signers]
      properties:
        signingMode:
          type: string
          enum: [ordered, parallel]
        signers:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [email, displayName, routingOrder]
            properties:
              signerId:
                type: string
                format: uuid
              email:
                type: string
              displayName:
                type: string
              routingOrder:
                type: integer
    ReplaceFieldsRequest:
      type: object
      additionalProperties: false
      required: [fields]
      properties:
        fields:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [signerId, type, pageNumber, x, y, width, height]
            properties:
              signerId:
                type: string
                format: uuid
              type:
                type: string
                enum: [signature, initials, date_signed, signer_name]
              pageNumber:
                type: integer
              x:
                type: number
              y:
                type: number
              width:
                type: number
              height:
                type: number
              required:
                type: boolean
    SendDocumentRequest:
      type: object
      additionalProperties: false
      properties:
        expiresAt:
          type: string
          format: date-time
          nullable: true
    SendDocumentResponse:
      allOf:
        - $ref: '#/components/schemas/PublicDocument'
        - type: object
          additionalProperties: false
          required: [invitations]
          properties:
            invitations:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [signerId, sessionId, expiresAt, token]
                properties:
                  signerId:
                    type: string
                    format: uuid
                  sessionId:
                    type: string
                    format: uuid
                  expiresAt:
                    type: string
                    format: date-time
                  token:
                    type: string
                    nullable: true
    RotateSessionResponse:
      type: object
      additionalProperties: false
      required: [signerId, sessionId, expiresAt, token]
      properties:
        signerId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        token:
          type: string
...